deleted no longer working tests with little value
